cat_dog/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
matplotlib>=3.0.0
numpy>=1.20.0
tqdm>=4.0.0      
requests>=2.0.0